release: v0.2.0 - audit-fix + 7 promised-feature implementations
Closed all 45 audit findings from issues/issuse 1 kairos/ISSUES.md (6 Critical,
14 High, 16 Medium, 7 Low, 2 Info) AND shipped seven features that v0.1
documented but had not implemented.

7 promised-but-missing features now real
- KAI-004: .kairos/config.toml parser (env > file > defaults).
- KAI-005: WikiIndexer + populated wiki_index/wiki_relations write paths.
- KAI-006: Runner ABC with entry_points("kairos.runners") plugin discovery.
- KAI-011: real kairos doctor MCP probe (was hard-coded "ok").
- KAI-017: MCPLLMClient retries + exponential backoff (3 attempts, 0.5s base).
- KAI-020: optional kairos run --llm-rerank for selector tie-breaks.
- KAI-035: kairos feedback CLI command (was dead schema).

Other v0.2 surfaces
- KAI-001: kairos init --force writes .bak.<UTC-timestamp> backups before overwrite.
- KAI-002: hash-aware re-ingest; same bytes = no-op, different = name_2.md.
- KAI-003: slug collision counter for source pages.
- KAI-007: console.print(..., markup=False) so [[wikilinks]] render correctly.
- KAI-008: real selector telemetry (selected_by, selector_score) propagated.
- KAI-010: lint findings tagged provenance=local|llm; LLM-orphans dropped.
- KAI-012: kairos query --save-to-wiki adds date + counter to filenames.
- KAI-016: YAML "Norway problem" boolean coerce in PageFrontmatter parser.
- KAI-018: KAIROS_DB_HOME env var honored.
- KAI-022: SQLite WAL + busy_timeout pragmas for concurrent runs.
- KAI-024: preserve created date on re-ingest.
- KAI-028: segment-match slug scoring (no more substring false-positives).
- KAI-029: top-N tie-break promotes default technique within 0.05 of leader.
- KAI-034: drop RunResult.trace field; --json reads trace from disk.
- KAI-042: duration_ms is float (REAL in SQLite).
- KAI-044: 5MB per-file guard in RAG runner.

Tests + CI
- 88 tests pass (was 48). +14 new test files. ruff + mypy strict clean
  across src/ AND tests/. New tests/integration/ uses httpx.MockTransport.
- CI gains pip-audit (KAI-013), pytest-cov, README demo drift check (KAI-014),
  mypy on tests/ (KAI-041).

Docs
- docs/UPGRADING.md: v0.1 -> v0.2 migration walkthrough (zero breaking changes).
- docs/architecture.md rewritten to match v0.2 reality; aspirational items
  moved under "v0.3+ surfaces" section.
- README "What's new in v0.2" quick-nav section.
- docs/launch/v0.2.md: audit-story LinkedIn launch kit (post + 6-slide
  carousel + engagement plan).

Removed
- [postgres] extra in pyproject.toml (deferred to v0.3).
- pytest-asyncio dev dep (no async tests).
- internal _parse_plan_json fallback_slug param (always unused).

Backwards compatibility
- All v0.1 CLI args, env vars, and on-disk formats preserved.
- v0.1.1 -> v0.2.0 in-place upgrade. First run after upgrade backfills
  wiki_index from FS.

+++ src/kairos/memory/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS runs (
     selected_by  TEXT NOT NULL,
     selector_score REAL,
     status       TEXT NOT NULL,
-    duration_ms  INTEGER,
+    duration_ms  REAL,
     cost_tokens  INTEGER,
     answer_path  TEXT,
     trace_path   TEXT,
